Handle validations errors from Castle's API
CastleDevise will "deny" requests that did not pass validation errors.
This is not ideal and will be changed in the future. A rationale for
this is that a missing or invalid "request_token" most likely means
the request has been made by a bot. At the moment our Ruby SDK gives
us no way of differentiating a missing/invalid request_token from other
validation errors.

Once the SDK has been updated, we will handle these two cases
differently:
- missing/invalid request_token will lead in a "deny" action
- other validation errors will be logged

We might also provide the ability to customize error handling in the
future.